refactor(cluster): split distributor service lifecycle support
